+++ Source/Gui/Controls/LogListBox.cpp
@@ -153,7 +153,7 @@ LRESULT CLogListBox::OnMeasureItem(UINT uMsg, WPARAM wParam, LPARAM lParam,BOOL&
     CString str;
     lpmis->itemWidth = ItemWidth;
     lpmis->itemHeight = LLB_VertMargin + item->TitleHeight + LLB_VertDivider + item->TextHeight + (item->InfoHeight?(LLB_VertDivider + item->InfoHeight):0) + LLB_VertMargin+2;
-    lpmis->itemHeight = max(lpmis->itemHeight, dpiScaleY_ * 70 );
+    lpmis->itemHeight = max(lpmis->itemHeight, static_cast<UINT>(dpiScaleY_ * 70) );
     lpmis->itemHeight = min(254 , lpmis->itemHeight);
 
     ReleaseDC(dc);
